+++ src/services/get_balance.ts
